+++ Entities/Actor.cpp
@@ -139,7 +139,7 @@ void Actor::Clear() {
     m_DamageMultiplier = 1.0F;
 
 	m_Organic = false;
-	m_Robotic = false;
+	m_Mechanical = false;
 }
 
 
@@ -289,7 +289,7 @@ int Actor::Create(const Actor &reference)
     m_TeamBlockState = reference.m_TeamBlockState;
 
 	m_Organic = reference.m_Organic;
-	m_Robotic = reference.m_Robotic;
+	m_Mechanical = reference.m_Mechanical;
 
     return 0;
 }
@@ -388,8 +388,8 @@ int Actor::ReadProperty(const std::string_view &propName, Reader &reader)
         m_AIMode = static_cast<AIMode>(mode);
 	} else if (propName == "Organic") {
 		reader >> m_Organic;
-	} else if (propName == "Robotic") {
-		reader >> m_Robotic;
+	} else if (propName == "Mechanical") {
+		reader >> m_Mechanical;
 	} else
         return MOSRotating::ReadProperty(propName, reader);
 
@@ -468,7 +468,7 @@ int Actor::Save(Writer &writer) const
     writer << m_AIMode;
 
 	writer.NewPropertyWithValue("Organic", m_Organic);
-	writer.NewPropertyWithValue("Robotic", m_Robotic);
+	writer.NewPropertyWithValue("Mechanical", m_Mechanical);
 
     return 0;
 }
@@ -1939,8 +1939,9 @@ void Actor::DrawHUD(BITMAP *pTargetBitmap, const Vector &targetPos, int whichScr
     }
 
     // AI Mode team roster HUD lines
-    if (/*m_Controller.IsState(PIE_MENU_ACTIVE) || */m_Controller.IsState(ACTOR_NEXT_PREP) || m_Controller.IsState(ACTOR_PREV_PREP))
-    {
+	if (g_ActivityMan.GetActivity()->GetViewState(g_ActivityMan.GetActivity()->PlayerOfScreen(whichScreen)) == Activity::ViewState::ActorSelect && g_SceneMan.ShortestDistance(m_Pos, g_SceneMan.GetScrollTarget(whichScreen), g_SceneMan.SceneWrapsX()).GetMagnitude() < 100) {
+		draw_sprite(pTargetBitmap, GetAIModeIcon(), cpuPos.m_X - 6, cpuPos.m_Y - 6);
+	} else if (m_Controller.IsState(ACTOR_NEXT_PREP) || m_Controller.IsState(ACTOR_PREV_PREP)) {
         int prevColor = m_Controller.IsState(ACTOR_PREV_PREP) ? 122 : (m_Team == Activity::TeamOne ? 13 : 147);
         int nextColor = m_Controller.IsState(ACTOR_NEXT_PREP) ? 122 : (m_Team == Activity::TeamOne ? 13 : 147);
         int prevSpacing = m_Controller.IsState(ACTOR_PREV_PREP) ? 3 : 9;